requirements.txt
reportlab==4.1.0
pillow==10.3.0
rlPyCairo==0.3.0
openpyxl==3.1.2